refactor(gossip): split the byte-entry codec out of the cache policy
bytes.go held two concerns and tripped Overall Code Complexity once the
admission split added functions to it. The stored record's format is the one
thing a cached entry must agree on across deploys, and its reader and writer
only make sense read together, so unwrapEntry, storeEntry, the marker bytes
and the scratch pool move to entry.go. What is left in bytes.go is the policy
that decides when to read, when to serve stale, and who pays for a refill.

No suppressions.
